refactor: madsim as a feature, delete real mod
