+++ core/src/layers/madsim.rs
@@ -315,3 +315,41 @@ struct ReadResponse {
 }
 
 struct WriteResponse {}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use crate::{services, EntryMode, Operator};
+    use madsim::{net::NetSim, runtime::Handle, time::sleep};
+    use std::time::Duration;
+
+    #[madsim::test]
+    async fn test_madsim_layer() {
+        let handle = Handle::current();
+        let ip1 = "10.0.0.1".parse().unwrap();
+        let ip2 = "10.0.0.2".parse().unwrap();
+        let sim_server_socket = "10.0.0.1:2379".parse().unwrap();
+        let server = handle.create_node().name("server").ip(ip1).build();
+        let client = handle.create_node().name("client").ip(ip2).build();
+
+        server.spawn(async move {
+            SimServer::serve(sim_server_socket).await.unwrap();
+        });
+        sleep(Duration::from_secs(1)).await;
+
+        let handle = client.spawn(async move {
+            let mut builder = services::Fs::default();
+            builder.root(".");
+            let op = Operator::new(builder)
+                .unwrap()
+                .layer(MadsimLayer::new(sim_server_socket))
+                .finish();
+
+            let path = "hello.txt";
+            let data = "Hello, World!";
+            op.write(path, data).await.unwrap();
+            assert_eq!(data.as_bytes(), op.read(path).await.unwrap());
+        });
+        handle.await.unwrap();
+    }
+}
